requirements.txt
langchain
langchain-community
langchain-groq
chromadb
streamlit
pandas
numpy
python-dotenv
sentence-transformers
langchain_huggingface
